Changes to Header/Sidebar/Main Content
Logo moved 10px to the right, just so it looks nicer on the page. Changed the font for the navigation bar since it is easier to read. Edited the navigation bar for each of our pages. And added links to them. I added the sidebar as well, it's still a bit faulty for the moment. Then I added the main content on the home page so far.